general logging improvements
removes file-based logging altogether. A user of the client is expected
to import into a project with file logging configured at the root
handler, or to import qbertconfig and configure the logger to include a
file handler if desired.

fixes issue with creating a kubeconfig file in the current directory

suppress Warnings from keystoneclient